requirements.txt
aiohttp==3.8.1
apscheduler==3.9.1
PyYAML==6.0
pytest==7.3.1
flake8==6.0.0
watchdog==2.1.9 